refactor(campaigns): decompose campaign detail page into focused sub-components
- Created campaign tab components:
  - CampaignOverviewTab.tsx (298L) - resources, missions, roster
  - CampaignAuditTab.tsx (256L) - timeline and event history
  - CampaignStatusHelpers.tsx (288L) - status display utilities
- Reduced src/pages/gameplay/campaigns/[id].tsx from 1,078L to 207L
- All 72 characterization tests passing
- Largest component: 298L (< 400L limit)
