+++ src/components/campaign/index.ts
@@ -4,6 +4,24 @@ export type {
   BlockingAction,
   BlockingActionType,
 } from './AdvanceDayButton';
+export { CampaignAuditTab } from './CampaignAuditTab';
+export type { CampaignAuditTabProps } from './CampaignAuditTab';
+export { CampaignOverviewTab } from './CampaignOverviewTab';
+export type { CampaignOverviewTabProps } from './CampaignOverviewTab';
+export {
+  getStatusColor,
+  getStatusLabel,
+  CampaignDetailSkeleton,
+  CampaignNotFound,
+  CampaignTabBar,
+  ResourcesCard,
+  MissionHistory,
+} from './CampaignStatusHelpers';
+export type {
+  CampaignTabBarProps,
+  ResourcesCardProps,
+  MissionHistoryProps,
+} from './CampaignStatusHelpers';
 export { MissionTreeView } from './MissionTreeView';
 export { RosterStateDisplay } from './RosterStateDisplay';
 export { StarmapDisplay, FACTION_COLORS } from './StarmapDisplay';
